+++ DP.h
@@ -17,6 +17,6 @@ IntGraph make_binary(const IntGraph &G, int r);
 /// Trying geometric discretization.
 map<int, edge_variables, std::greater<>> DP_arb_modified(const IntGraph &H, int r, double epsilon, int S, int R_of_MST);
 
-map<int, DP_type, std::greater<>> DP_bidirectional(const IntGraph &H, int r, double epsilon, long S, long R_of_MST);
+map<int, DP_type, std::greater<>> DP_bidirectional(const IntGraph &H, int r, double epsilon, long long S, long long R_of_MST);
 
 #endif //GRAPH_VERSIONING_DP_H
